requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
mysql-connector-python
joblibpython-docx
